backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
qrcode[pil]==7.4.2
Pillow==10.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
zipfile36==0.1.3
geoip2==4.7.0
user-agents==2.2.0